refactor: share the route-test app builder via conftest
The browse and search route tests each carried a near-identical `app`
fixture: same template folder, Babel, path-config context processor,
url_escape global, endpoint stubs, and `client` fixture. Only the
blueprint and the stubbed endpoint list actually differed, so those become
the two arguments to make_route_app().

Drop the PIKARAOKE_BASE_PATH and PIKARAOKE_SOCKETIO_PATH config keys the
browse test set. Nothing reads them: app.py's context processor is their
only consumer and both test apps register their own.

The werkzeug.__version__ shim stays duplicated. It appears in seven test
modules and hoisting it would mean touching five files this branch does
not otherwise change.
